+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This changelog records changes to stable releases since 1.50.2. "TBA" changes he
 
 ## Nightly (only)
 
+Nothing, yet
+
+## v1.84 (October 2023)
+
+- feat: improve event listener breakpoints view ([#1853](https://github.com/microsoft/vscode-js-debug/issues/1853))
 - fix: envFiles variables appending rather than replacing in attach ([vscode#1935510](https://github.com/microsoft/vscode/issues/1935510))
 - fix: cache-bust sourcemaps if the underlying content changes ([#1803](https://github.com/microsoft/vscode-js-debug/issues/1803))
 - fix: make source map renames scope-aware